Refactor implemenations into there own spots
