+++ server/src/main/java/org/opensearch/identity/noop/NoopAuthenticationManger.java
@@ -1,0 +1,7 @@
+package org.opensearch.identity.noop;
+
+import org.opensearch.identity.AuthenticationManager;
+
+public class NoopAuthenticationManger implements AuthenticationManager {
+    
+}
