+++ server/src/main/java/org/opensearch/identity/shiro/ShiroAuthenticationManager.java
@@ -1,4 +1,4 @@
-package org.opensearch.identity;
+package org.opensearch.identity.shiro;
 
 import java.security.Principal;
 import java.util.concurrent.Callable;
